feat(user rate): score format implementation
- added score formats for AniList API
- added score format change option in SettingsScreen
- adjusted saveServiceUserRate function to work with different score formats
- fixed an issue with the slider in UserRateBottomSheet that caused certain values to be skipped when changing the slider's value

+++ app/src/main/java/com/example/shikiflow/domain/model/tracks/UserRateData.kt
@@ -93,7 +93,7 @@ data class SaveUserRate(
     val mediaId: Int,
     val malId: Int?,
     val userStatus: UserRateStatus,
-    val score: Int = 0,
+    val score: Float = 0f,
     val progress: Int = 0,
     val progressVolumes: Int? = null,
     val repeat: Int = 0
